requirements.txt
apache-airflow
apache-airflow-providers-docker
pandas
requests
sqlalchemy
psycopg2-binary
loguru
